% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deid_dems_single_row_per_student}
\alias{deid_dems_single_row_per_student}
\title{deid_dems_single_row_per_student.}
\format{A data frame with eleven variables: \code{fake_student_id}, \code{gender},
  \code{race}, \code{race_descr}, \code{migrant_ind}, \code{homeless_ind},
  \code{sped_ind}, \code{sped_code}, \code{sped_descr}, \code{low_income_ind},
   and \code{ell_status}.}
\usage{
deid_dems_single_row_per_student
}
\description{
Demographic data of students. One student per row of data.
Short dataset with only 11 students and 11 variables.
}
\keyword{datasets}

